clause/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
google-genai>=1.3.0
httpx>=0.27.0
pypdf>=5.1.0
